Adjust version checks for Base deprecation overwriting
